LF happiness limit and apple changes
LFs now have a limit of happiness from 0 to 100. Benefits come from
higher happiness as stated in previous updates.

LFs will now only chase after an apple IF their personal happiness is
less than 100. If it is 100, the apple will be ignored because the hap.
can't be raised anymore. However, if a LF with 100 happiness comes in
contact with an apple, the apple will still be consumed. Just without
any benefits.